pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "hr-rag-assistant"
version = "0.1.0"
description = "A lightweight RAG assistant for HR policy Q&A with citations."
requires-python = ">=3.10,<3.13"
authors = [{name = "HR RAG Assistant Maintainers"}]
readme = "README.md"
dependencies = [
  "openai>=1.6.0",
  "python-dotenv>=1.0.1",
  "faiss-cpu>=1.8.0",
  "pydantic>=2.6.0",
  "streamlit>=1.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.5.0",
  "mypy>=1.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

